Complete rewrite of the low-level interface
Manage to have a low level interface (in `AT.jl`) which mimics as closely as
possible the calls to the SDK C functions.  After calling `using AT`, all
functions and constants are prefixed by `AT.` The conventions are that all
low-level functions return a status and expected output values if any.  The
status is an instance of `AT.Status` with embedded symbolic function name and
integer code.  For instance:

    using AT
    AT.GetEnumCount(handle, feature) -> status, count

At a higher level, the `check` method behave like a filter which takes
the result of the call to a low level method and throw an error if the
returned status indicates a failure and pass the remaining arguments to the
caller otherwise.