Move all type definitions from deps.jl to AT.jl
These types were all manually (not automatically) converted from
definitions found in <atcore.h> so they need not be in deps.jl
which is generated by the package manager.

The only remaining "automatic" defintions are #define'd constants
which can be parsed by Julia.  A compiler should ne longer be needed.